feat: use server data for dashboards
