+++ src/app/dashboard/scouting/page.tsx
@@ -19,142 +19,8 @@ import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem } from "
 import { cn } from "@/lib/utils"
 import { Separator } from "@/components/ui/separator"
 
-// Datos de ejemplo - en producción vendrían de la API
-const jugadoresScouteados = [
-  { 
-    id: "1", 
-    nombre: "Alejandro", 
-    apellido: "Martínez", 
-    equipo: "CD Leganés", 
-    añoNacimiento: 2012,
-    demarcacion: "Centrocampista",
-    lateralidad: "Derecha",
-    altura: "Alto",
-    complexion: "Atlético",
-    capacidadFisica: 4,
-    capacidadTecnica: 5,
-    capacidadTactica: 4,
-    capacidadDefensiva: 3,
-    mental: 4,
-    portero: 0,
-    propuesta: "FICHAR",
-    dificultad: "POSIBLE",
-    observaciones: "Excelente visión de juego y técnica. Destaca por su capacidad de pase y control.",
-    fechaScouting: "2025-03-15",
-    entrenador: "Carlos Pérez",
-    equipoCDSA: "Infantil A",
-    categoria: "1ª Infantil",
-    historial: [
-      {
-        fechaScouting: "2024-05-20",
-        equipo: "EF Alcobendas",
-        capacidadFisica: 3,
-        capacidadTecnica: 4,
-        capacidadTactica: 3,
-        capacidadDefensiva: 3,
-        mental: 3,
-        propuesta: "SEGUIMIENTO",
-        observaciones: "Jugador con potencial pero necesita mejorar físicamente."
-      }
-    ]
-  },
-  { 
-    id: "2", 
-    nombre: "Daniel", 
-    apellido: "García", 
-    equipo: "Rayo Vallecano", 
-    añoNacimiento: 2013,
-    demarcacion: "Delantero",
-    lateralidad: "Izquierda",
-    altura: "Altura media",
-    complexion: "Atlético",
-    capacidadFisica: 5,
-    capacidadTecnica: 4,
-    capacidadTactica: 3,
-    capacidadDefensiva: 2,
-    mental: 5,
-    portero: 0,
-    propuesta: "FICHAR",
-    dificultad: "COMPLICADO",
-    observaciones: "Delantero muy rápido y con gran definición. Destaca por su velocidad y capacidad goleadora.",
-    fechaScouting: "2025-02-28",
-    entrenador: "Miguel López",
-    equipoCDSA: "Alevín A",
-    categoria: "1ª Alevín"
-  },
-  { 
-    id: "3", 
-    nombre: "Pablo", 
-    apellido: "Rodríguez", 
-    equipo: "Atlético de Madrid", 
-    añoNacimiento: 2011,
-    demarcacion: "Defensa",
-    lateralidad: "Derecha",
-    altura: "Alto",
-    complexion: "Fuerte",
-    capacidadFisica: 4,
-    capacidadTecnica: 3,
-    capacidadTactica: 4,
-    capacidadDefensiva: 5,
-    mental: 4,
-    portero: 0,
-    propuesta: "NO FICHAR",
-    dificultad: "IMPOSIBLE",
-    observaciones: "Defensa central con gran capacidad física y buen juego aéreo. Pertenece a la cantera del Atlético.",
-    fechaScouting: "2025-01-15",
-    entrenador: "Laura Sánchez",
-    equipoCDSA: "Cadete B",
-    categoria: "2ª Cadete"
-  },
-  { 
-    id: "4", 
-    nombre: "Mario", 
-    apellido: "Fernández", 
-    equipo: "Getafe CF", 
-    añoNacimiento: 2014,
-    demarcacion: "Portero",
-    lateralidad: "Derecha",
-    altura: "Alto",
-    complexion: "Atlético",
-    capacidadFisica: 3,
-    capacidadTecnica: 4,
-    capacidadTactica: 4,
-    capacidadDefensiva: 3,
-    mental: 5,
-    portero: 5,
-    propuesta: "FICHAR",
-    dificultad: "POSIBLE",
-    observaciones: "Portero con grandes reflejos y buen juego con los pies. Destaca por su colocación y seguridad.",
-    fechaScouting: "2025-03-05",
-    entrenador: "Pedro Martín",
-    equipoCDSA: "Benjamín A",
-    categoria: "1ª Benjamín"
-  },
-  { 
-    id: "5", 
-    nombre: "Javier", 
-    apellido: "López", 
-    equipo: "AD Alcorcón", 
-    añoNacimiento: 2012,
-    demarcacion: "Centrocampista",
-    lateralidad: "Ambas",
-    altura: "Altura media",
-    complexion: "Delgado",
-    capacidadFisica: 3,
-    capacidadTecnica: 5,
-    capacidadTactica: 5,
-    capacidadDefensiva: 3,
-    mental: 4,
-    portero: 0,
-    propuesta: "FICHAR",
-    dificultad: "POSIBLE",
-    observaciones: "Centrocampista muy técnico y con gran visión de juego. Ambidiestro y con gran capacidad de pase.",
-    fechaScouting: "2025-02-10",
-    entrenador: "Carlos Pérez",
-    equipoCDSA: "Infantil A",
-    categoria: "1ª Infantil"
-  }
-]
+// Datos de ejemplo - en producción vendrán de la API
+const jugadoresScouteados: any[] = []
 
 // Opciones para los selectores
 const demarcaciones = ["Portero", "Defensa", "Centrocampista", "Delantero"]
